chore: refactoring Stats
