+++ TelnyxRTC/Telnyx/WebRTC/Peer.swift
@@ -22,9 +22,9 @@ class Peer : NSObject {
     private let VIDEO_TRACK_ID = "video0"
     //TODO: REMOVE THIS FOR V1
     private let VIDEO_DEMO_LOCAL_VIDEO = "local_video_streaming.mp4"
-    private var gatheredICECandidates:[String] = []
-    var socket:Socket? 
-    private let timeStamp = Timestamp()
+    private var gatheredICECandidates: [String] = []
+    var socket: Socket?
+
 
     private let mediaConstrains = [kRTCMediaConstraintsOfferToReceiveAudio: kRTCMediaConstraintsValueTrue, kRTCMediaConstraintsOfferToReceiveVideo: kRTCMediaConstraintsValueFalse]
 
@@ -39,7 +39,7 @@ class Peer : NSObject {
     private var videoCapturer: RTCVideoCapturer?
     private var localVideoTrack: RTCVideoTrack?
     private var remoteVideoTrack: RTCVideoTrack?
-    private var callLegID: String?
+    internal var callLegID: String?
 
     //Data channel
     private var localDataChannel: RTCDataChannel?
@@ -48,6 +48,21 @@ class Peer : NSObject {
     //ICE negotiation
     private var negotiationTimer: Timer?
     private var negotiationEnded: Bool = false
+
+    // WEBRTC STATS
+    internal var timer: DispatchSourceTimer?
+    internal let timeStamp = Timestamp()
+    internal var statsEvent = [String: Any]()
+    internal var inboundStats = [Any]()
+    internal var outBoundStats = [Any]()
+    internal var statsData = [String: Any]()
+    internal var audio = [String: [Any]]()
+    internal var candidatePairs =  [Any]()
+    internal let CANDIDATE_PAIR_LIMIT = 5
+    internal var debugStatsId = UUID.init()
+    internal var debugReportStarted = false
+    internal var isDebugStats = false
+    
 
     // The `RTCPeerConnectionFactory` is in charge of creating new RTCPeerConnection instances.
     // A new RTCPeerConnection should be created every new call, but the factory is shared.
@@ -201,108 +216,6 @@ class Peer : NSObject {
             })
         }
     }
-    
-    private var timer: DispatchSourceTimer?
-
-    private var statsEvent = [String: Any]()
-    private var inboundStats = [Any]()
-    private var outBoundStats = [Any]()
-    private var statsData = [String: Any]()
-    private var audio = [String: [Any]]()
-    private var candidatePairs =  [Any]()
-    private let CANDIDATE_PAIR_LIMIT = 5
-    private var debugStatsId = UUID.init()
-    private var debugReportStarted = false
-    private var isDebugStats = false
-
-        func startTimer() {
-            isDebugStats = true
-            let queue = DispatchQueue.main
-            timer = DispatchSource.makeTimerSource(queue: queue)
-            timer?.schedule(deadline: .now(), repeating: 2.0)
-            timer?.setEventHandler { [weak self] in
-                self?.executeTask()
-            }
-            timer?.resume()
-        }
-
-        func stopTimer() {
-            statsData["audio"] = audio
-            statsEvent["data"] = statsData
-            statsEvent.printJson()
-            timer?.cancel()
-            timer = nil
-            sendStatsType(id: debugStatsId, type: StatsType.STOP_STARTS.rawValue)
-            debugReportStarted = false
-            isDebugStats = false
-        }
-    
-    /// To receive INVITE message after Push Noficiation is Received. Send attachCall Command
-    fileprivate func sendStats(id:UUID,data:[String:Any]) {
-        Logger.log.e(message: "TxClient:: Sending Stats")
-        let statsMessage = StatsMessage(reportID: id.uuidString.lowercased(), reportData: data)
-        self.socket?.sendMessage(message: statsMessage.encode())
-    }
-    
-    fileprivate func sendStatsType(id:UUID,type:String) {
-        Logger.log.e(message: "TxClient:: Sending Stats \(type)")
-        let statsMessage = InitiateOrStopStats(type: type, reportID: id.uuidString.lowercased())
-        self.socket?.sendMessage(message: statsMessage.encode())
-    }
-    
-    
-
-        private func executeTask() {
-            print("Task executed at \(Date())")
-            
-            
-            if(!debugReportStarted){
-                debugStatsId = UUID.init()
-                sendStatsType(id: debugStatsId, type: StatsType.START_STARTS.rawValue)
-                debugReportStarted = true
-            }
-          
-            statsEvent["event"] = "stats"
-            statsEvent["tag"] = "stats"
-            statsEvent["peerId"] = "stats"
-            statsEvent["connectionId"] = self.callLegID ?? ""
-            statsEvent["timeTaken"] = 1
-            
-            
-
-            self.connection?.statistics(completionHandler: { reports in
-                reports.statistics.forEach { report in
-                    if(report.value.type == "inbound-rtp") {
-                        //Logger.log.i(message: "Peer:: ICE negotiation updated. Report New: \(report.values)")
-                        self.inboundStats.append(report.value.values)
-                    }
-                    if(report.value.type == "outbound-rtp") {
-                        //Logger.log.i(message: "Peer:: ICE negotiation updated. Report New: \(report.values)")
-                        self.outBoundStats.append(report.value.values)
-                    }
-                    if(report.value.type == "candidate-pair" && self.candidatePairs.count < self.CANDIDATE_PAIR_LIMIT) {
-                        //Logger.log.i(message: "Peer:: ICE negotiation updated. Report New: \(report.values)")
-                        self.candidatePairs.append(report.value.values)
-                    }
-                }
-            })
-            audio["outbound"] = outBoundStats
-            audio["inbound"] = inboundStats
-            statsData["audio"] = audio
-            statsEvent["data"] = statsData
-            statsEvent["timestamp"] = timeStamp.getTimestamp()
-
-            if(inboundStats.count > 0 && outBoundStats.count > 0 && candidatePairs.count > 0){
-                inboundStats.removeAll()
-                outBoundStats.removeAll()
-                candidatePairs.removeAll()
-                statsData.removeAll()
-                audio.removeAll()
-                self.sendStats(id: debugStatsId, data: statsEvent)
-            }
-
-        }
-
 
     /**
      This code should be started when the first ICE candidate is created.
@@ -344,8 +257,8 @@ class Peer : NSObject {
     /// Close connection and release resources
     func dispose() {
         Logger.log.i(message: "Peer:: dispose()")
-        if(isDebugStats){
-            stopTimer()
+        if isDebugStats {
+            self.stopTimer()
         }
         
         //This should release all the connection resources
@@ -519,60 +432,3 @@ extension Peer : RTCPeerConnectionDelegate {
         Logger.log.i(message: "Peer:: connection didOpen RTCDataChannel: \(dataChannel)")
     }
 }
-
-// MARK: - Dictionary
-extension Dictionary {
-
-    var json: String {
-        let invalidJson = "Not a valid JSON"
-        do {
-            let jsonData = try JSONSerialization.data(withJSONObject: self, options: .prettyPrinted)
-            return String(bytes: jsonData, encoding: String.Encoding.utf8) ?? invalidJson
-        } catch {
-            return invalidJson
-        }
-    }
-
-    func printJson() {
-        print(json)
-    }
-
-}
-
-// MARK: - Stats
-
-private let PROTOCOL_VERSION: String = "2.0"
-
-
-
-enum StatsType : String  {
-    case STOP_STARTS = "debug_report_stop"
-    case START_STARTS = "debug_report_start"
-}
-
-class InitiateOrStopStats {
-    
-    private var jsonMessage: [String: Any] = [String: Any]()
-    let jsonrpc = PROTOCOL_VERSION
-    var id: String = UUID.init().uuidString.lowercased()
-    
-    init(type:String,reportID:String){
-        self.jsonMessage = [String: Any]()
-        self.jsonMessage["jsonrpc"] = self.jsonrpc
-        self.jsonMessage["id"] = self.id
-        self.jsonMessage["debug_report_version"] = 1
-        self.jsonMessage["type"] = type
-        self.jsonMessage["debug_report_id"] = reportID
-    }
-    
-    func encode() -> String? {
-        guard let jsonData = try? JSONSerialization.data(withJSONObject: jsonMessage, options: []),
-              let jsonString = String(data: jsonData, encoding: .utf8) else {
-            Logger.log.e(message: "Message:: encode() error")
-            return nil
-        }
-        return jsonString
-    }
-}
-
-
